// apps/game-client/src/lib/api/generated/main/model/create-notification-dto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { CreateNotificationDtoNpc } from './create-notification-dto-npc';

export interface CreateNotificationDto {
  /** @maxLength 500 */
  message?: string;
  npc?: CreateNotificationDtoNpc;
  /**
     * @minItems 1
     * @maxItems 10
     */
  guildIds: string[];
  /**
     * @minLength 1
     * @maxLength 50
     */
  world: string;
  isGatheringParty?: boolean;
}