// apps/game-client/src/lib/api/generated/main/model/create-party-gathering-dto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { CreatePartyGatheringDtoCharacter } from './create-party-gathering-dto-character.js';

export interface CreatePartyGatheringDto {
  /**
     * @minItems 1
     * @maxItems 10
     * @items.maxLength 50
     */
  guildIds: string[];
  /**
     * @minLength 1
     * @maxLength 50
     */
  world: string;
  character: CreatePartyGatheringDtoCharacter;
  /** @maxLength 200 */
  description?: string;
  /**
     * @minimum 1
     * @maximum 500
     */
  minLvl?: number;
  /**
     * @minimum 1
     * @maximum 500
     */
  maxLvl?: number;
}