// apps/game-client/src/lib/api/generated/main/model/create-volunteer-dto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { CreateVolunteerDtoCharacter } from './create-volunteer-dto-character';

export interface CreateVolunteerDto {
  /**
     * @minLength 1
     * @maxLength 50
     */
  world: string;
  /**
     * @minLength 1
     * @maxLength 20
     */
  targetDiscordId: string;
  character: CreateVolunteerDtoCharacter;
}